use `u32` when dealing with OTA sizes

+++ src/consts.rs
@@ -19,4 +19,4 @@ pub const WRITE_TIMEOUT: Duration = Duration::from_secs(4);
 pub const NOTIFICATION_MAX_LEN: usize = 1024;
 
 /// Default chunk size for upgrade parts.
-pub const UPDATE_PART_SIZE: usize = 1024; // 1 kB
+pub const UPDATE_PART_SIZE: u32 = 1024; // 1 kB
